Path fixes, also fixed an old bug
For some reason, I hard-coded the indent level when de-indenting (in the parser). It would only work for one indent, but not for two.

Refactored things in `compile`, going through the path to Windows compatibility. The main issue will be the install commands (because of installation directories), but that'll hopefully be done in the next week.